+++ mayan/apps/dynamic_search/api_views.py
@@ -4,10 +4,9 @@ from rest_framework.exceptions import ParseError
 
 from mayan.apps.rest_api import generics
 
-from .classes import SearchModel
+from .classes import SearchBackend, SearchModel
 from .mixins import SearchModelAPIMixin
 from .serializers import SearchModelSerializer
-from .runtime import search_backend
 
 
 class APISearchView(SearchModelAPIMixin, generics.ListAPIView):
@@ -25,7 +24,7 @@ class APISearchView(SearchModelAPIMixin, generics.ListAPIView):
             self.mayan_object_permissions = {'GET': (search_model.permission,)}
 
         try:
-            queryset = search_backend.search(
+            queryset = SearchBackend.get_instance().search(
                 search_model=search_model,
                 query_string=self.request.GET, user=self.request.user
             )
@@ -63,7 +62,7 @@ class APIAdvancedSearchView(SearchModelAPIMixin, generics.ListAPIView):
             global_and_search = False
 
         try:
-            queryset = search_backend.search(
+            queryset = SearchBackend.get_instance().search(
                 global_and_search=global_and_search,
                 query_string=self.request.GET,
                 search_model=self.search_model, user=self.request.user