+++ mayan/apps/dynamic_search/apps.py
@@ -1,9 +1,14 @@
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.apps import MayanAppConfig
-from mayan.apps.common.menus import menu_facet, menu_secondary
+from mayan.apps.common.menus import menu_facet, menu_secondary, menu_tools
 
-from .links import link_search, link_search_advanced, link_search_again
+from .classes import SearchModel
+from .links import (
+    link_search, link_search_advanced, link_search_again,
+    link_search_backend_reindex
+)
+from .tasks import task_index_instance  # NOQA - Force task registration
 
 
 class DynamicSearchApp(MayanAppConfig):
@@ -17,6 +22,8 @@ class DynamicSearchApp(MayanAppConfig):
     def ready(self):
         super(DynamicSearchApp, self).ready()
 
+        SearchModel.initialize()
+
         menu_facet.bind_links(
             links=(link_search, link_search_advanced),
             sources=(
@@ -25,4 +32,7 @@ class DynamicSearchApp(MayanAppConfig):
         )
         menu_secondary.bind_links(
             links=(link_search_again,), sources=('search:results',)
+        )
+        menu_tools.bind_links(
+            links=(link_search_backend_reindex,),
         )