+++ mayan/apps/dynamic_search/classes.py
@@ -1,18 +1,67 @@
+from collections import Iterable
 import logging
 
 from django.apps import apps
+from django.db.models.signals import post_save, pre_delete
 from django.utils.encoding import force_text
+from django.utils.functional import cached_property
 from django.utils.module_loading import import_string
 from django.utils.translation import ugettext as _
 
+from mayan.apps.common.exceptions import ResolverPipelineError
 from mayan.apps.common.literals import LIST_MODE_CHOICE_LIST
+from mayan.apps.common.utils import (
+    ResolverPipelineModelAttribute, get_related_field
+)
 
+from .settings import (
+    setting_search_backend, setting_search_backend_arguments
+)
 logger = logging.getLogger(name=__name__)
 
 
 class SearchBackend:
-    def search(self, global_and_search, search_model, query_string, user):
+    @staticmethod
+    def get_instance():
+        return import_string(dotted_path=setting_search_backend.value)(
+            **setting_search_backend_arguments.value
+        )
+
+    def __init__(self, **kwargs):
+        self.kwargs = kwargs
+
+    def _search(self, global_and_search, search_model, query_string, user):
         raise NotImplementedError
+
+    def deindex_instance(self, instance):
+        raise NotImplementedError
+
+    def index_instance(self, instance):
+        raise NotImplementedError
+
+    def search(self, search_model, query_string, user, global_and_search=False):
+        AccessControlList = apps.get_model(
+            app_label='acls', model_name='AccessControlList'
+        )
+
+        # Clean up the query_string
+        # The original query_string is immutable, create a new
+        # mutable copy
+        query_string = query_string.copy()
+        query_string.pop('_match_all', None)
+
+        queryset = self._search(
+            global_and_search=global_and_search, search_model=search_model,
+            query_string=query_string, user=user
+        )
+
+        if search_model.permission:
+            queryset = AccessControlList.objects.restrict_queryset(
+                permission=search_model.permission, queryset=queryset,
+                user=user
+            )
+
+        return queryset
 
 
 class SearchField:
@@ -31,9 +80,51 @@ class SearchField:
     def get_model(self):
         return self.search_model.model
 
+    def get_model_field(self):
+        return get_related_field(
+            model=self.get_model(), related_field_name=self.field
+        )
+
 
 class SearchModel:
+    _model_search_relationships = {}
     _registry = {}
+
+    @staticmethod
+    def flatten_list(value):
+        if isinstance(value, (str, bytes)):
+            yield value
+        else:
+            for item in value:
+                if isinstance(item, Iterable) and not isinstance(item, (str, bytes)):
+                    yield from SearchModel.flatten_list(value=item)
+                else:
+                    yield item
+
+    @staticmethod
+    def function_return_same(value):
+        return value
+
+    @staticmethod
+    def initialize():
+        # Hide a circular import
+        from .handlers import (
+            handler_factory_deindex_instance, handler_index_instance
+        )
+
+        for search_model in SearchModel.all():
+            post_save.connect(
+                dispatch_uid='search_handler_index_instance',
+                receiver=handler_index_instance,
+            )
+            pre_delete.connect(
+                dispatch_uid='search_handler_deindex_instance_{}'.format(search_model),
+                receiver=handler_factory_deindex_instance(search_model=search_model),
+                sender=search_model.model,
+                weak=False
+            )
+
+            search_model._initialize()
 
     @classmethod
     def all(cls):
@@ -54,6 +145,10 @@ class SearchModel:
 
         return result
 
+    @classmethod
+    def get_for_model(cls, instance):
+        return cls.get(name=instance._meta.label)
+
     def __init__(
         self, app_label, model_name, serializer_path, label=None,
         list_mode=None, permission=None, queryset=None
@@ -69,8 +164,26 @@ class SearchModel:
         self.queryset = queryset
         self.__class__._registry[self.get_full_name()] = self
 
+    def __repr__(self):
+        return '<{}: {}>'.format(
+            self.__class__.__name__, self.label
+        )
+
     def __str__(self):
         return force_text(self.label)
+
+    def _initialize(self):
+        for search_field in self.search_fields:
+            related_model = get_related_field(
+                model=self.model, related_field_name=search_field.field
+            ).model
+
+            if related_model != self.model:
+                self.__class__._model_search_relationships.setdefault(self.model, set())
+                self.__class__._model_search_relationships[self.model].add(related_model)
+
+                self.__class__._model_search_relationships.setdefault(related_model, set())
+                self.__class__._model_search_relationships[related_model].add(self.model)
 
     def add_model_field(self, *args, **kwargs):
         """
@@ -104,18 +217,49 @@ class SearchModel:
         except KeyError:
             raise KeyError('No search field named: %s' % full_name)
 
-    @property
+    @cached_property
     def label(self):
         if not self._label:
             self._label = self.model._meta.verbose_name
         return self._label
 
-    @property
+    @cached_property
     def model(self):
         if not self._model:
             self._model = apps.get_model(self.app_label, self.model_name)
         return self._model
 
-    @property
+    @cached_property
     def pk(self):
         return self.get_full_name()
+
+    def sieve(self, field_map, instance):
+        """
+        Method that receives an instance and a field map dictionary
+        consisting of attribute names and transformations to apply.
+        Returns a dictionary of the instance values with their respective
+        transformations. Makes it easy to pre process an instance before
+        indexing it.
+        """
+        result = {}
+        for field in field_map:
+            try:
+                value = ResolverPipelineModelAttribute.resolve(
+                    attribute=field, obj=instance
+                )
+                try:
+                    value = list(SearchModel.flatten_list(value))
+                    if value == [None]:
+                        value = None
+                    else:
+                        value = ''.join(value)
+                except TypeError:
+                    """Value is not a list"""
+            except ResolverPipelineError:
+                """Not fatal"""
+            else:
+                result[field] = field_map[field].get(
+                    'transformation', SearchModel.function_return_same
+                )(value)
+
+        return result