+++ mayan/apps/dynamic_search/literals.py
@@ -1,0 +1,1 @@
+TASK_RETRY_DELAY = 5