+++ mayan/apps/dynamic_search/views.py
@@ -1,16 +1,21 @@
 import logging
 
+from django.contrib import messages
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic.base import RedirectView
 
-from mayan.apps.common.generics import SimpleView, SingleObjectListView
+from mayan.apps.common.generics import (
+    ConfirmView, SimpleView, SingleObjectListView
+)
 from mayan.apps.common.literals import LIST_MODE_CHOICE_ITEM
 
+from .classes import SearchBackend, SearchModel
 from .forms import SearchForm, AdvancedSearchForm
 from .icons import icon_search_submit
 from .mixins import SearchModelMixin
-from .runtime import search_backend
+from .permissions import permission_search_tools
+from .tasks import task_index_search_model
 
 logger = logging.getLogger(name=__name__)
 
@@ -45,13 +50,44 @@ class ResultsView(SearchModelMixin, SingleObjectListView):
             else:
                 global_and_search = False
 
-            queryset = search_backend.search(
+            queryset = SearchBackend.get_instance().search(
                 global_and_search=global_and_search,
                 search_model=self.search_model,
                 query_string=self.request.GET, user=self.request.user
             )
 
             return queryset
+
+
+class SearchAgainView(RedirectView):
+    pattern_name = 'search:search_advanced'
+    query_string = True
+
+
+class SearchBackendReindexView(ConfirmView):
+    extra_context = {
+        'title': _('Reindex search backend'),
+        'subtitle': _(
+            'This tool is required only for some search backends.'
+        ),
+    }
+    view_permission = permission_search_tools
+
+    def get_post_action_redirect(self):
+        return reverse(viewname='common:tools_list')
+
+    def view_action(self):
+        for search_model in SearchModel.all():
+            task_index_search_model.apply_async(
+                kwargs={
+                    'search_model_full_name': search_model.get_full_name(),
+                }
+            )
+
+        messages.success(
+            message=_('Search backend reindexing queued.'),
+            request=self.request
+        )
 
 
 class SearchView(SearchModelMixin, SimpleView):
@@ -89,8 +125,3 @@ class AdvancedSearchView(SearchView):
         return AdvancedSearchForm(
             data=self.request.GET, search_model=self.get_search_model()
         )
-
-
-class SearchAgainView(RedirectView):
-    pattern_name = 'search:search_advanced'
-    query_string = True